root_orchestrator/system-manager-python/requirements.txt
Flask~=2.1.1
marshmallow~=3.15.0
flask-smorest
Flask_PyMongo
flask_socketio
eventlet~=0.33.0
PyYAML~=6.0
requests~=2.27.1
Cerberus
prometheus-client
jsonschema~=4.4.0
flask_cors
PyPi
flask_jwt_extended
flask_restful
pymongo~=4.1.1
werkzeug==2.0.3
MarkupSafe~=2.1.1
flask_swagger_ui
mongomock==4.1.2
paho-mqtt==1.6.1